[DAGE-15] design of (not definitive) test for docker-compose
