+++ rre-dataset-generator/tests/integration/resources/good_solr_config.yaml
