+++ rre-dataset-generator/tests/integration/test_elasticsearch.py
@@ -12,14 +12,14 @@ def test_index_exists(search_url):
     """Check that the 'testcore' exists."""
     # r = requests.get(search_url + "testcore")
     # assert r.ok, f"Index 'testcore' not found: {r.text}"
-    assert True
+    assert False
 
 def test_index_has_docs(search_url):
     """Check that there are documents in the index."""
     # r = requests.get(search_url + "testcore/_search", params={"size": 0})
     # assert r.ok, r.text
     # assert r.json()["hits"]["total"]["value"] > 0
-    assert True
+    assert False
 
 @pytest.mark.parametrize("field", ["_id", "title"])
 def test_docs_have_field(search_url, field):
@@ -28,11 +28,11 @@ def test_docs_have_field(search_url, field):
     # assert r.ok, r.text
     # doc = r.json()["hits"]["hits"][0]["_source"]
     # assert field in doc, f"Field '{field}' not in document: {doc}"
-    assert True
+    assert False
 
 
 def test_search_engine_fetch(search_url):
     # engine = ElasticsearchSearchEngine(search_url)
     # docs   = engine.fetch_for_query_generation(None, 3, ["title", "description"])
     # assert len(docs) == 3 and all(d.id and d.fields for d in docs)
-    assert True
+    assert False
