+++ rre-dataset-generator/tests/integration/test_vespa.py
@@ -11,22 +11,22 @@ if os.environ.get("PYTEST_CURRENT_TEST", "").endswith("test_vespa.py::"):
 
 def test_index_exists(search_url):
     """Check that the 'test-index' exists."""
-    assert True
+    assert False
 
 
 def test_index_has_docs(search_url):
     """Check that there are documents in the index."""
-    assert True
+    assert False
 
 
 @pytest.mark.parametrize("field", ["_id", "title"])
 def test_docs_have_field(search_url, field):
     """Verify that a given field exists in a sample doc."""
-    assert True
+    assert False
 
 
 def test_search_engine_fetch(search_url):
     # engine = VespaSearchEngine(search_url)
     # docs   = engine.fetch_for_query_generation(None, 3, ["title", "description"])
     # assert len(docs) == 3 and all(d.id and d.fields for d in docs)
-    assert True
+    assert False
